+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Recursive completion on Graphic handle, struct, handle , class (properties, methods).
 - adding links between documents about mex and supported compilers.
 - GitHub CI for macOS Sonoma (Apple Silicon) support.
+- `Export to ...` context menu for console and text editor as pdf.
+- `CTRL + Mouse wheel` or `CTRL + +/-` to zoom in/out on console, editor, help.
 
 ### Fixed